requirements.txt
# Core Framework (versões mais recentes - Nov 2025)
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.19
jinja2==3.1.5

# LangGraph e LangChain (versões mais recentes)
langgraph==1.0.3
langchain==0.3.12
langchain-core==0.3.23
langchain-anthropic==0.3.8
langchain-openai==0.3.9
langchain-google-genai==2.0.8

# Data Validation
pydantic==2.10.3
pydantic-settings==2.7.0

# AI/ML Providers (versões mais recentes)
anthropic==0.42.0
openai==1.58.1
google-generativeai==0.8.5

# Utilities
httpx==0.28.1
python-dotenv==1.0.1
aiofiles==24.1.0

# Image Processing (opcional)
pillow==11.0.0
